+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
   -------------------------------------------------------------------
 ## [Unreleased]
+- Added cli warnings for deprecated positional arguments and environmental variables.
 
 ## [1.16.0] 2021-03-16
 - Added support for LZMA `*.xz` compression on file output. To use this feature, Specify an output path ending with `.xz`.